refactor: move calculateCarnetData to a new file
